requirements.txt
pandas~=1.3.5
requests~=2.27.1
python-dotenv~=0.19.2
twarc~=2.8.3
openpyxl~=3.0.9